[skip ci] get opConfig opts working
