+++ asset/tsconfig.json
@@ -5,7 +5,8 @@
         "module": "ESNext",
         "outDir": "dist",
         "rootDir": "src",
-        "esModuleInterop": true
+        "esModuleInterop": true,
+        "types": ["node"]
     },
     "include": ["src", "test"]
 }
