+++ tsconfig.json
@@ -20,9 +20,8 @@
         "sourceMap": true,
         "ignoreDeprecations": "5.0",
         "disableReferencedProjectLoad": true,
-        "typeRoots": [
-            "./node_modules/@types"
-        ]
+        "typeRoots": ["./node_modules/@types"],
+        "types": ["jest", "node"]
     },
     "include": ["test"],
     "references": [
